requirements.txt
pyTelegramBotAPI
Flask
psycopg2-binary
qrcode
pillow
